Revert "Addding Final BackEnd"
This reverts commit 22b6d7213aea911c7670d27c730f08cf2e93989c.

+++ project_spring_boot_backend/src/main/java/com/app/entities/SearchRestByCuisineId.java
@@ -17,10 +17,10 @@ import lombok.Setter;
 public class SearchRestByCuisineId implements Serializable{
 	@Column(name="cuisine_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long cuisineId ;
+	private Integer cuisineId ;
 	
 	@Column(name="rest_id")
-	private Long restId ;
+	private Integer restId ;
 	
 	@Override
     public boolean equals(Object o) {
